refactor: Move shift methods into shift class
