+++ spec/shift_spec.rb
@@ -1,4 +1,5 @@
 require './lib/shift'
+require './lib/enigma'
 
 RSpec.describe Shift do
   before(:each) do 
@@ -7,5 +8,10 @@ RSpec.describe Shift do
 
   it 'exists' do
     expect(@shift).to be_a Shift
+  end
+
+  it 'can shift a number' do
+    expect(@shift.shift_number(10, 10)).to eq 20
+    expect(@shift.shift_number(34, 2)).to eq 9
   end
 end
