+++ spec/spec_helper.rb
@@ -2,3 +2,4 @@ require 'simplecov'
 
 SimpleCov.start
 require_relative 'enigma_spec'
+require_relative 'shift_spec'
